+++ scripts/lib/bashy-core/stderr-messages.sh
@@ -37,7 +37,7 @@ _stderr_errorEnabled=1
 _stderr_infoEnabled=1
 
 # Whether progress messages are enabled.
-_stderr_progressEnabled=0
+_stderr_progressEnabled=1
 
 
 #
@@ -50,18 +50,47 @@ function error-msg {
     _stderr_print-handler '_stderr_errorEnabled' '_stderr_anyErrors' "$@"
 }
 
-# Prints an info message to stderr, if such are enabled. **Note:** Info
-# messages are _enabled_ by default.
+# Prints an info or warning message to stderr, if such are enabled. **Note:**
+# Info messages are _enabled_ by default.
 function info-msg {
     _stderr_print-handler '_stderr_infoEnabled' '' "$@"
 }
 
 # Prints a progress message to stderr, if such are enabled. **Note:** Progress
-# messages are _disabled_ by default.
+# messages are _enabled_ by default.
 function progress-msg {
     _stderr_print-handler '_stderr_progressEnabled' '' "$@"
 }
 
+# Prints an option which reasonably approximates the current stderr settings.
+# Specifically, it prints `--verbose=<level>` with the most detailed <level>
+# currently enabled.
+function stderr-opt {
+    printf '%s' '--verbose='
+
+    if (( _stderr_progressEnabled )); then
+        echo 'all'
+    elif (( _stderr_infoEnabled )); then
+        echo 'warn'
+    elif (( _stderr_errorEnabled )); then
+        echo 'error'
+    else
+        echo 'none'
+    fi
+}
+
+# Adds the usual stderr-control options.
+# * `--verbose=<level>` -- Indicates which kinds of messages will pass. From
+#   least to most:
+#   * `none` -- None.
+#   * `error` -- Just errors.
+#   * `warn` -- Warning and informational messages.
+#   * `all` -- Everything.
+# * `--quiet` / `-q` -- Same as `--verbose=none`
+function usual-stderr-args {
+    opt-value --call=_stderr_verbose --enum='none error warn all' verbose
+    opt-action --call='{ _stderr_verbose none }' quiet/q
+}
 
 #
 # Internal functions
@@ -130,4 +159,32 @@ function _stderr_print-handler {
         # `printf` to avoid option-parsing weirdness with `echo`.
         printf 1>&2 '%s\n' "$*"
     fi
+}
+
+# Handles the options `--quiet` and `--verbose`.
+function _stderr_verbose {
+    local level="$1"
+
+    case "${level}" in
+        none)
+            error-msg --disable
+            info-msg --disable
+            progress-msg --disable
+            ;;
+        error)
+            error-msg --enable
+            info-msg --disable
+            progress-msg --disable
+            ;;
+        warn)
+            error-msg --enable
+            info-msg --enable
+            progress-msg --disable
+            ;;
+        all)
+            error-msg --enable
+            info-msg --enable
+            progress-msg --enable
+            ;;
+    esac
 }