+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Notable changes:
   * `ubik`:
     * Finds functions from the library, not just commands (scripts).
     * Smarter error reporting.
+  * `progress-msg` defaults to enabled.
 * `bashy-basics`:
   * New script `timey secs`. Subsumes and deprecates `timey now-secs`.
   * New helper functions `jbash-array` and `jpostproc`.