+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Notable changes:
 * `bashy-basics`:
   * New script `timey secs`. Subsumes and deprecates `timey now-secs`.
   * New helper functions `jbash-array` and `jpostproc`.
+  * Renamed commands `json-*` to just `j*`.
 
 ### v2.1 -- 2023-05-12